[errorprone] Prevent int expression may overflow
see https://errorprone.info/bugpattern/IntLongMath

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/PackObjectSizeIndexV1.java
@@ -122,7 +122,7 @@ class PackObjectSizeIndexV1 implements PackObjectSizeIndex {
 
 	@Override
 	public long getObjectCount() {
-		return positions24.size() + positions32.length;
+		return (long) positions24.size() + positions32.length;
 	}
 
 	@Override
